+++ src/app/app.module.ts
@@ -1,7 +1,7 @@
 import {ActsAsUser} from './services/acts-as-user.service';
 import {AgmCoreModule} from 'angular2-google-maps/core';
 import {ApiCall} from './services/api-call.service';
-import {ApiErrorsComponent} from './components/forms/api-errors.component';
+import {ApiErrorsComponent} from './components/form-errors/api-errors/api-errors.component';
 import {AppComponent} from './app.component';
 import {AppRoutingModule} from './app.routing.module';
 import {AppTranslateModule} from './app.translate.module';
@@ -38,8 +38,8 @@ import {HomeComponent} from './views/home/home.component';
 import {HomeTopBannerComponent} from './components/home/home-top-banner.component';
 import {HttpModule} from '@angular/http';
 import {InfoMessageComponent} from './components/messages/info-message/info-message.component';
-import {InputErrorComponent} from './components/forms/input-error.component';
-import {InputErrorsComponent} from './components/forms/input-errors.component';
+import {InputErrorComponent} from './components/form-errors/input-error/input-error.component';
+import {InputErrorsComponent} from './components/form-errors/input-errors/input-errors.component';
 import {JARoutes} from './routes/ja-routes';
 import {JobCardComponent} from './components/cards/job-card/job-card.component';
 import {JobDetailsComponent} from './views/job-details/job-details.component';
